Move MUC functions to utils
